added ASSERT assembly test
