+++ tests/test_Assembly.py
@@ -1853,39 +1853,6 @@ RETURN %
 
 
 class TestAssert(TestCase):
-    def test_assert(self):
-        @apply_operations
-        def target(x):
-            assembly(
-                """
-ASSERT $x
-"""
-            )
-
-        dis.dis(target)
-
-        target(1)
-        self.assertRaises(AssertionError, lambda: target(0))
-
-    def test_assert_with_message(self):
-        @apply_operations
-        def target(x):
-            assembly(
-                """
-ASSERT $x \"Test Message\"
-"""
-            )
-
-        target(1)
-        self.assertRaises(AssertionError, lambda: target(0))
-
-        try:
-            target(0)
-        except AssertionError as e:
-            self.assertEqual(e.args, ("Test Message",))
-        else:
-            self.assertTrue(False)
-
     def test_assert_static(self):
         @apply_operations
         def target():
